pyproject.toml
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ranking-system"
version = "0.1.0"
description = "Multi-stage LLM-augmented ranking system with retrieval and feedback simulation"
requires-python = ">=3.11"

dependencies = [
    # Retrieval
    "faiss-cpu>=1.7.4",
    "rank-bm25>=0.2.2",
    "sentence-transformers>=2.7.0",

    # Ranking
    "lightgbm>=4.3.0",
    "scikit-learn>=1.4.0",
    "numpy>=1.26.0",
    "pandas>=2.2.0",

    # LLM
    "openai>=1.30.0",
    "anthropic>=0.28.0",
    "tiktoken>=0.7.0",        # token counting for cost tracking

    # Config
    "hydra-core>=1.3.2",
    "omegaconf>=2.3.0",

    # Evaluation & tracking
    "mlflow>=2.13.0",
    "scipy>=1.13.0",

    # Utils
    "python-dotenv>=1.0.0",
    "rich>=13.7.0",            # beautiful terminal output
    "tqdm>=4.66.0",
    "loguru>=0.7.2",           # structured logging, better than logging module
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.0",             # linting + formatting, replaces black+flake8
    "ipykernel>=6.29.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"